+++ app/Http/Controllers/ProductManufacturerImageController.php
@@ -3,8 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Models\ProductManufacturerImage;
 use App\Services\StorageHandlers\DynamicStorageHandler;
+use App\Models\ProductManufacturerImage;
+use App\Models\Product;
 
 class ProductManufacturerImageController extends Controller
 {
@@ -15,6 +16,10 @@ class ProductManufacturerImageController extends Controller
         ]);
 
         $productId = $request->input('productId');
+
+        $product = Product::find($productId);
+
+        $this->authorize('create', [ProductManufacturerImage::class, $product]);
 
         $uploadInfo = DynamicStorageHandler::store($request->file('image'), 'product-manufacturer-images', 'public');
         
@@ -33,6 +38,10 @@ class ProductManufacturerImageController extends Controller
 
     public function destroy($productManufacturerImageId)
     {
+        $productManufacturerImage = ProductManufacturerImage::with('product')->find($productManufacturerImageId);
+
+        $this->authorize('delete', $productManufacturerImage);
+
         ProductManufacturerImage::where('id', $productManufacturerImageId)->delete();
 
         session(['success_message' => 'Deleted successfully.']);
